web: Fix echo behavior of WebActionDatapoint+WebDisplayDatapoint
This change fixes the behaviour for interactive web widgets which are
WebDisplayDatapoint and WebActionDatapoint and have their
convert_from_ws value and/or convert_to_ws_value overridden in a that
does not simply do a type conversion. An example is the ValueButton,
which converts all values from the Websocket to the preset value but
always sends 'True' to the websocket when the value is equal to the
preset value.