web: Add automatic page reload on server restart
We add the WebServer object id to each page's source code, so the
JavaScript can transmit it upon websocket connect. This way, restarts of
the server between page load and websocket (re)connect can be detected
and the client can be asked to reload the page in this situation.

This mechanism is required to make sure the WebConnector's object ids in
the page HTML source and the server process match, to ensure that
routing of websocket messages works.