Fix deprecation warnings for asyncio.get_event_loop()
In future Python versions, asyncio.get_event_loop() will not create a
new event loop as required. As of Python 3.10 it issues a deprecation
warning when it has to.

This commit changes get_event_loop().run_until_complete() to
asyncio.run(), where possible. However, sometimes we need the event loop
to stay alive after running a coroutine, so we need to ensure having an
event loop manually, similar to how asyncio.run() does.